@@ -9,8 +9,6 @@
 		"lint": "vue-cli-service lint"
 	},
 	"dependencies": {
-		"@types/js-cookie": "^2.2.0",
-		"@types/nprogress": "0.0.29",
 		"axios": "^0.18.0",
 		"element-ui": "^2.4.11",
 		"font-awesome": "^4.7.0",
@@ -26,15 +24,17 @@
 		"vuex": "^3.0.1"
 	},
 	"devDependencies": {
+		"@types/js-cookie": "^2.2.0",
+		"@types/nprogress": "0.0.29",
 		"@vue/cli-plugin-babel": "^3.2.0",
 		"@vue/cli-plugin-eslint": "^3.2.0",
 		"@vue/cli-plugin-typescript": "^3.2.0",
 		"@vue/cli-service": "^3.2.0",
-		"@vue/eslint-config-prettier": "^4.0.0",
+		"@vue/eslint-config-prettier": "^4.0.1",
 		"@vue/eslint-config-typescript": "^3.2.0",
 		"babel-eslint": "^10.0.1",
-		"eslint": "^5.8.0",
-		"eslint-plugin-vue": "^5.0.0-0",
+		"eslint": "^5.11.1",
+		"eslint-plugin-vue": "^5.0.0",
 		"json-loader": "^0.5.7",
 		"node-sass": "^4.9.0",
 		"sass-loader": "^7.0.1",
@@ -48,28 +48,12 @@
 		},
 		"extends": [
 			"plugin:vue/essential",
+			"@vue/prettier",
 			"@vue/typescript"
 		],
-		"rules": {
-			"quotes": [
-				"error",
-				"single"
-			],
-			"indent": [
-				"error",
-				"tab",
-				{
-					"SwitchCase": 1
-				}
-			]
-		},
+		"rules": {},
 		"parserOptions": {
-			"parser": "typescript-eslint-parser",
-			"ecmaFeatures": {
-				"parser": "babel-eslint",
-				"jsx": false,
-				"globalReturn": true
-			}
+			"parser": "typescript-eslint-parser"
 		}
 	},
 	"postcss": {
@@ -81,5 +65,8 @@
 		"> 1%",
 		"last 2 versions",
 		"not ie <= 8"
-	]
+	],
+	"prettier": {
+		"trailingComma": "es5"
+	}
 }
