@@ -1,7 +1,6 @@
 <template>
   <div>
-    <svg
-      :class="{'is-active':isActive}"
+    <svg :class="{'is-active':isActive}"
       t="1492500959545"
       class="hamburger"
       style=""
@@ -28,17 +27,17 @@
 
 <script>
 export default {
-	name: 'Hamburger',
+	name: "Hamburger",
 	props: {
 		isActive: {
 			type: Boolean,
-			default: false
+			default: false,
 		},
 		toggleClick: {
 			type: Function,
-			default: null
-		}
-	}
+			default: null,
+		},
+	},
 };
 </script>
 
@@ -49,7 +48,7 @@ export default {
 	width: 20px;
 	height: 20px;
 	transform: rotate(90deg);
-	transition: .38s;
+	transition: 0.38s;
 	transform-origin: 50% 50%;
 }
 .hamburger.is-active {
