@@ -8,19 +8,19 @@
 const padding = 15; // tag's padding
 
 export default {
-	name: 'ScrollPane',
-	data () {
+	name: "ScrollPane",
+	data() {
 		return {
-			left: 0
+			left: 0,
 		};
 	},
 	methods: {
-		handleScroll (e) {
+		handleScroll(e) {
 			const eventDelta = e.wheelDelta || -e.deltaY * 40;
 			const $scrollWrapper = this.$refs.scrollContainer.$refs.wrap;
 			$scrollWrapper.scrollLeft += eventDelta / 4;
 		},
-		moveToTarget ($target) {
+		moveToTarget($target) {
 			const $container = this.$refs.scrollContainer.$el;
 			const $containerWidth = $container.offsetWidth;
 			const $scrollWrapper = this.$refs.scrollContainer.$refs.wrap;
@@ -28,29 +28,30 @@ export default {
 			const $targetWidth = $target.offsetWidth;
 			if ($targetLeft > $containerWidth) {
 				// tag in the right
-				$scrollWrapper.scrollLeft = $targetLeft - $containerWidth + $targetWidth + padding;
+				$scrollWrapper.scrollLeft =
+					$targetLeft - $containerWidth + $targetWidth + padding;
 			} else {
 				// tag in the left
 				$scrollWrapper.scrollLeft = $targetLeft - padding;
 			}
-		}
-	}
+		},
+	},
 };
 </script>
 
 <style rel="stylesheet/scss" lang="scss" scoped>
 .scroll-container {
-  white-space: nowrap;
-  position: relative;
-  overflow: hidden;
-  width: 100%;
-  /deep/ {
-    .el-scrollbar__bar {
-      bottom: 0px;
-    }
-    .el-scrollbar__wrap {
-      height: 49px;
-    }
-  }
+	white-space: nowrap;
+	position: relative;
+	overflow: hidden;
+	width: 100%;
+	/deep/ {
+		.el-scrollbar__bar {
+			bottom: 0px;
+		}
+		.el-scrollbar__wrap {
+			height: 49px;
+		}
+	}
 }
 </style>
