@@ -1,4 +1,4 @@
-declare module '.config' {
+declare module ".config" {
 	const config: JSON;
 	export default config;
 }
