@@ -1,30 +1,33 @@
-import api from '@/api';
-import router from '@/router';
-import { removeToken } from '@/utils/auth';
+import api from "@/api";
+import router from "@/router";
+import { removeToken } from "@/utils/auth";
 
 export default {
 	ToggleSideBar: ({ commit }) => {
-		commit('TOGGLE_SIDEBAR');
+		commit("TOGGLE_SIDEBAR");
 	},
-	CloseSideBar ({ commit }, { withoutAnimation }) {
-		commit('CLOSE_SIDEBAR', withoutAnimation);
+	CloseSideBar({ commit }, { withoutAnimation }) {
+		commit("CLOSE_SIDEBAR", withoutAnimation);
 	},
-	OpenSideBar ({ commit }, { withoutAnimation }) {
-		commit('OPEN_SIDEBAR', withoutAnimation);
+	OpenSideBar({ commit }, { withoutAnimation }) {
+		commit("OPEN_SIDEBAR", withoutAnimation);
 	},
-	ToggleDevice ({ commit }, device) {
-		commit('TOGGLE_DEVICE', device);
+	ToggleDevice({ commit }, device) {
+		commit("TOGGLE_DEVICE", device);
 	},
-	GetProfile ({ commit }) {
-		return api.get('dis/me').then(result => {
-			commit('GET_PROFILE', result);
+	GetProfile({ commit }) {
+		return api.get("dis/me").then(result => {
+			commit("GET_PROFILE", result);
 		});
 	},
-	LogOut ({ commit }) {
-		api.del('dis/access-tokens').then(() => {
-			commit('GET_PROFILE', {});
+	LogOut({ commit }) {
+		api.del("dis/access-tokens").then(() => {
+			commit("GET_PROFILE", {});
 			removeToken();
-			router.push({ path: '/login', query: { redirect: router.app.$route.fullPath } });
+			router.push({
+				path: "/login",
+				query: { redirect: router.app.$route.fullPath },
+			});
 		});
-	}
+	},
 };
