@@ -1,11 +1,11 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
-import * as getters from './getters';
-import state from './state';
-import mutations from './mutations';
-import actions from './actions';
-import tagsView from './modules/tagsView';
-import checkLogin from './modules/checkLogin';
+import Vue from "vue";
+import Vuex from "vuex";
+import * as getters from "./getters";
+import state from "./state";
+import mutations from "./mutations";
+import actions from "./actions";
+import tagsView from "./modules/tagsView";
+import checkLogin from "./modules/checkLogin";
 
 Vue.use(Vuex);
 
@@ -17,6 +17,6 @@ const store = new Vuex.Store({
 	state,
 	getters,
 	mutations,
-	actions
+	actions,
 });
 export default store;
