@@ -1,4 +1,4 @@
-import api from '@/api';
+import api from "@/api";
 
 const checkLogin = {
 	state: {
@@ -6,23 +6,23 @@ const checkLogin = {
 		user: undefined,
 	},
 	mutations: {
-		SET_LOGIN (state, data) {
+		SET_LOGIN(state, data) {
 			state.isLogin = data.status;
 			state.user = data.user;
-		}
+		},
 	},
 	actions: {
-		checkLogin ({ commit }) {
-			return api.get('dis/access-tokens').then(result => {
+		checkLogin({ commit }) {
+			return api.get("dis/access-tokens").then(result => {
 				if (result.code == 2000) {
 					result.status = true;
-					commit('SET_LOGIN', result);
+					commit("SET_LOGIN", result);
 				} else {
-					commit('SET_LOGIN', result);
+					commit("SET_LOGIN", result);
 				}
 			});
-		}
-	}
+		},
+	},
 };
 
 export default checkLogin;
