@@ -1,17 +1,17 @@
-import router from '@/router';
-import nprogress from 'nprogress';
-import 'nprogress/nprogress.css';
-import { getToken } from './auth';
+import router from "@/router";
+import nprogress from "nprogress";
+import "nprogress/nprogress.css";
+import { getToken } from "./auth";
 
 nprogress.inc(0.2);
-nprogress.configure({ easing: 'ease', speed: 500, showSpinner: false });
+nprogress.configure({ easing: "ease", speed: 500, showSpinner: false });
 
-router.beforeEach((to:any, from:any, next:any) => {
+router.beforeEach((to: any, from: any, next: any) => {
 	nprogress.start();
-	if (!getToken() && to.path != '/login') {
+	if (!getToken() && to.path != "/login") {
 		next({
-			path: '/login',
-			query: { redirect: to.fullPath }
+			path: "/login",
+			query: { redirect: to.fullPath },
 		});
 	} else {
 		next();
