@@ -17,44 +17,44 @@
 </template>
 
 <script lang="ts">
-import WeatherInfo from './components/WeatherInfo.vue';
+import WeatherInfo from "./components/WeatherInfo.vue";
 
 export default {
-	name: 'Index',
+	name: "Index",
 	components: {
-		WeatherInfo
+		WeatherInfo,
 	},
 	data() {
 		return {
 			panelData: {},
-			detail: require('@/assets/bduss.png')
+			detail: require("@/assets/bduss.png"),
 		};
 	},
 	computed: {},
 	methods: {},
-	created() {}
+	created() {},
 };
 </script>
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
 <style lang="scss" scoped>
-  .dashboard-editor-container {
-    position: relative;
-    padding: 32px;
-    background-color: rgb(240, 242, 245);
-    .chart-wrapper {
-      background: #fff;
-      padding: 16px 16px 0;
-      margin-bottom: 32px;
-    }
-  }
-  @media only screen and (max-width: 768px) {
-    .dashboard-editor-container {
-      padding: 16px;
-      .chart-wrapper {
-        padding: 8px 8px 0;
-        margin-bottom: 16px;
-      }
-    }
-  }
+.dashboard-editor-container {
+	position: relative;
+	padding: 32px;
+	background-color: rgb(240, 242, 245);
+	.chart-wrapper {
+		background: #fff;
+		padding: 16px 16px 0;
+		margin-bottom: 32px;
+	}
+}
+@media only screen and (max-width: 768px) {
+	.dashboard-editor-container {
+		padding: 16px;
+		.chart-wrapper {
+			padding: 8px 8px 0;
+			margin-bottom: 16px;
+		}
+	}
+}
 </style>
