@@ -15,11 +15,11 @@
 </template>
 
 <script>
-import { Navbar, Sidebar, AppMain, TagsView } from './components';
-import ResizeMixin from './mixin/ResizeHandler';
+import { Navbar, Sidebar, AppMain, TagsView } from "./components";
+import ResizeMixin from "./mixin/ResizeHandler";
 
 export default {
-	name: 'Layout',
+	name: "Layout",
 	components: {
 		Navbar,
 		Sidebar,
@@ -28,48 +28,48 @@ export default {
 	},
 	mixins: [ResizeMixin],
 	computed: {
-		sidebar () {
+		sidebar() {
 			return this.$store.state.sidebar;
 		},
-		device () {
+		device() {
 			return this.$store.state.device;
 		},
-		classObj () {
+		classObj() {
 			return {
 				hideSidebar: !this.sidebar.opened,
 				openSidebar: this.sidebar.opened,
 				withoutAnimation: this.sidebar.withoutAnimation,
-				mobile: this.device === 'mobile'
+				mobile: this.device === "mobile",
 			};
-		}
+		},
 	},
 	methods: {
-		handleClickOutside () {
-			this.$store.dispatch('CloseSideBar', { withoutAnimation: false });
-		}
-	}
+		handleClickOutside() {
+			this.$store.dispatch("CloseSideBar", { withoutAnimation: false });
+		},
+	},
 };
 </script>
 
 <style rel="stylesheet/scss" lang="scss" scoped>
-	@import "src/styles/mixin.scss";
-	.app-wrapper {
-	  @include clearfix;
-	  position: relative;
-	  height: 100%;
-	  width: 100%;
-	  &.mobile.openSidebar {
-	    position: fixed;
-	    top: 0;
-	  }
+@import "src/styles/mixin.scss";
+.app-wrapper {
+	@include clearfix;
+	position: relative;
+	height: 100%;
+	width: 100%;
+	&.mobile.openSidebar {
+		position: fixed;
+		top: 0;
 	}
-	.drawer-bg {
+}
+.drawer-bg {
 	//   background: #000;
-	  opacity: 0.3;
-	  width: 100%;
-	  top: 0;
-	  height: 100%;
-	  position: absolute;
-	  z-index: 999;
-	}
+	opacity: 0.3;
+	width: 100%;
+	top: 0;
+	height: 100%;
+	position: absolute;
+	z-index: 999;
+}
 </style>
