@@ -10,20 +10,20 @@
 
 <script>
 export default {
-	name: 'AppMain',
+	name: "AppMain",
 	computed: {
 		// key() {
 		//   return this.$route.name !== undefined ? this.$route.name + +new Date() : this.$route + +new Date()
 		// }
-	}
+	},
 };
 </script>
 
 <style scoped>
 .app-main {
-  /*50 = navbar  */
-  min-height: calc(100vh - 50px);
-  position: relative;
-  overflow: hidden;
+	/*50 = navbar  */
+	min-height: calc(100vh - 50px);
+	position: relative;
+	overflow: hidden;
 }
 </style>
