@@ -30,35 +30,35 @@
 </template>
 
 <script>
-import path from 'path';
-import { validateURL } from '@/utils/validate';
-import Item from './Item';
+import path from "path";
+import { validateURL } from "@/utils/validate";
+import Item from "./Item";
 
 export default {
-	name: 'SidebarItem',
+	name: "SidebarItem",
 	components: { Item },
 	props: {
 		// route object
 		item: {
 			type: Object,
-			required: true
+			required: true,
 		},
 		isNest: {
 			type: Boolean,
-			default: false
+			default: false,
 		},
 		basePath: {
 			type: String,
-			default: ''
-		}
+			default: "",
+		},
 	},
-	data () {
+	data() {
 		return {
-			onlyOneChild: null
+			onlyOneChild: null,
 		};
 	},
 	methods: {
-		hasOneShowingChild (children, parent) {
+		hasOneShowingChild(children, parent) {
 			const showingChildren = children.filter(item => {
 				if (item.hidden) {
 					return false;
@@ -66,7 +66,6 @@ export default {
 				// Temp set(will be used if only has one showing child)
 				this.onlyOneChild = item;
 				return true;
-
 			});
 
 			// When there is only one child router, the child router is displayed by default
@@ -76,19 +75,23 @@ export default {
 
 			// Show parent if there are no child router to display
 			if (showingChildren.length === 0) {
-				this.onlyOneChild = { ...parent, path: '', noShowingChildren: true };
+				this.onlyOneChild = {
+					...parent,
+					path: "",
+					noShowingChildren: true,
+				};
 				return true;
 			}
 
 			return false;
 		},
-		resolvePath (routePath) {
+		resolvePath(routePath) {
 			return path.resolve(this.basePath, routePath);
 		},
-		isExternalLink (routePath) {
+		isExternalLink(routePath) {
 			return validateURL(routePath);
 		},
-		clickLink (routePath, e) {
+		clickLink(routePath, e) {
 			e.preventDefault();
 			if (!this.isExternalLink(routePath)) {
 				const path = this.resolvePath(routePath);
@@ -96,7 +99,7 @@ export default {
 			} else {
 				window.open(routePath);
 			}
-		}
-	}
+		},
+	},
 };
 </script>
