@@ -23,57 +23,60 @@
 </template>
 
 <script>
-import ScrollPane from '@/components/ScrollPane';
+import ScrollPane from "@/components/ScrollPane";
 
 export default {
 	components: { ScrollPane },
-	data () {
+	data() {
 		return {
 			visible: false,
 			top: 0,
 			left: 0,
-			selectedTag: {}
+			selectedTag: {},
 		};
 	},
 	computed: {
-		visitedViews () {
+		visitedViews() {
 			return this.$store.state.tagsView.visitedViews;
-		}
+		},
 	},
 	watch: {
-		$route () {
+		$route() {
 			this.addViewTags();
 			this.moveToCurrentTag();
 		},
-		visible (value) {
+		visible(value) {
 			if (value) {
-				document.body.addEventListener('click', this.closeMenu);
+				document.body.addEventListener("click", this.closeMenu);
 			} else {
-				document.body.removeEventListener('click', this.closeMenu);
+				document.body.removeEventListener("click", this.closeMenu);
 			}
-		}
+		},
 	},
-	mounted () {
+	mounted() {
 		this.addViewTags();
 	},
 	methods: {
-		generateRoute () {
+		generateRoute() {
 			if (this.$route.name) {
 				return this.$route;
 			}
 			return false;
 		},
-		isActive (route) {
-			return route.path === this.$route.path || route.path === `${this.$route.path}/`;
+		isActive(route) {
+			return (
+				route.path === this.$route.path ||
+				route.path === `${this.$route.path}/`
+			);
 		},
-		addViewTags () {
+		addViewTags() {
 			const route = this.generateRoute();
 			if (!route) {
 				return false;
 			}
-			this.$store.dispatch('addView', route);
+			this.$store.dispatch("addView", route);
 		},
-		moveToCurrentTag () {
+		moveToCurrentTag() {
 			const tags = this.$refs.tag;
 			this.$nextTick(() => {
 				for (const tag of tags) {
@@ -82,7 +85,10 @@ export default {
 
 						// when query is different then update
 						if (tag.to.fullPath !== this.$route.fullPath) {
-							this.$store.dispatch('updateVisitedView', this.$route);
+							this.$store.dispatch(
+								"updateVisitedView",
+								this.$route
+							);
 						}
 
 						break;
@@ -90,143 +96,145 @@ export default {
 				}
 			});
 		},
-		refreshSelectedTag (view) {
-			this.$store.dispatch('delCachedView', view).then(() => {
+		refreshSelectedTag(view) {
+			this.$store.dispatch("delCachedView", view).then(() => {
 				const { fullPath } = view;
 				this.$nextTick(() => {
 					this.$router.replace({
-						path: `/redirect${fullPath}`
+						path: `/redirect${fullPath}`,
 					});
 				});
 			});
 		},
-		closeSelectedTag (view) {
-			this.$store.dispatch('delView', view).then(({ visitedViews }) => {
+		closeSelectedTag(view) {
+			this.$store.dispatch("delView", view).then(({ visitedViews }) => {
 				if (this.isActive(view)) {
 					const latestView = visitedViews.slice(-1)[0];
 					if (latestView) {
 						this.$router.push(latestView);
 					} else {
-						this.$router.push('/');
+						this.$router.push("/");
 						this.addViewTags();
 					}
 				}
 			});
 		},
-		closeOthersTags () {
+		closeOthersTags() {
 			this.$router.push(this.selectedTag);
-			this.$store.dispatch('delOthersViews', this.selectedTag).then(() => {
-				this.moveToCurrentTag();
-			});
+			this.$store
+				.dispatch("delOthersViews", this.selectedTag)
+				.then(() => {
+					this.moveToCurrentTag();
+				});
 		},
-		closeAllTags () {
-			this.$store.dispatch('delAllViews');
-			this.$router.push('/');
+		closeAllTags() {
+			this.$store.dispatch("delAllViews");
+			this.$router.push("/");
 			this.addViewTags();
 		},
-		openMenu (tag, e) {
+		openMenu(tag, e) {
 			this.visible = true;
 			this.selectedTag = tag;
 			const offsetLeft = this.$el.getBoundingClientRect().left; // container margin left
 			this.left = e.clientX - offsetLeft + 15; // 15: margin right
 			this.top = e.clientY;
 		},
-		closeMenu () {
+		closeMenu() {
 			this.visible = false;
-		}
-	}
+		},
+	},
 };
 </script>
 
 <style rel="stylesheet/scss" lang="scss" scoped>
 .tags-view-container {
-  height: 34px;
-  width: 100%;
-  background: #fff;
-  border-bottom: 1px solid #d8dce5;
-  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, .12), 0 0 3px 0 rgba(0, 0, 0, .04);
-  .tags-view-wrapper {
-    .tags-view-item {
-      display: inline-block;
-      position: relative;
-      height: 26px;
-      line-height: 26px;
-      border: 1px solid #d8dce5;
-      color: #495060;
-      background: #fff;
-      padding: 0 8px;
-      font-size: 12px;
-      margin-left: 5px;
-      margin-top: 4px;
-      &:first-of-type {
-        margin-left: 15px;
-      }
-      &:last-of-type {
-        margin-right: 15px;
-      }
-      &.active {
-        background-color: #42b983;
-        color: #fff;
-        border-color: #42b983;
-        &::before {
-          content: '';
-          background: #fff;
-          display: inline-block;
-          width: 8px;
-          height: 8px;
-          border-radius: 50%;
-          position: relative;
-          margin-right: 2px;
-        }
-      }
-    }
-  }
-  .contextmenu {
-    margin: 0;
-    background: #fff;
-    z-index: 100;
-    position: absolute;
-    list-style-type: none;
-    padding: 5px 0;
-    border-radius: 4px;
-    font-size: 12px;
-    font-weight: 400;
-    color: #333;
-    box-shadow: 2px 2px 3px 0 rgba(0, 0, 0, .3);
-    li {
-      margin: 0;
-      padding: 7px 16px;
-      cursor: pointer;
-      &:hover {
-        background: #eee;
-      }
-    }
-  }
+	height: 34px;
+	width: 100%;
+	background: #fff;
+	border-bottom: 1px solid #d8dce5;
+	box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.12), 0 0 3px 0 rgba(0, 0, 0, 0.04);
+	.tags-view-wrapper {
+		.tags-view-item {
+			display: inline-block;
+			position: relative;
+			height: 26px;
+			line-height: 26px;
+			border: 1px solid #d8dce5;
+			color: #495060;
+			background: #fff;
+			padding: 0 8px;
+			font-size: 12px;
+			margin-left: 5px;
+			margin-top: 4px;
+			&:first-of-type {
+				margin-left: 15px;
+			}
+			&:last-of-type {
+				margin-right: 15px;
+			}
+			&.active {
+				background-color: #42b983;
+				color: #fff;
+				border-color: #42b983;
+				&::before {
+					content: "";
+					background: #fff;
+					display: inline-block;
+					width: 8px;
+					height: 8px;
+					border-radius: 50%;
+					position: relative;
+					margin-right: 2px;
+				}
+			}
+		}
+	}
+	.contextmenu {
+		margin: 0;
+		background: #fff;
+		z-index: 100;
+		position: absolute;
+		list-style-type: none;
+		padding: 5px 0;
+		border-radius: 4px;
+		font-size: 12px;
+		font-weight: 400;
+		color: #333;
+		box-shadow: 2px 2px 3px 0 rgba(0, 0, 0, 0.3);
+		li {
+			margin: 0;
+			padding: 7px 16px;
+			cursor: pointer;
+			&:hover {
+				background: #eee;
+			}
+		}
+	}
 }
 </style>
 
 <style rel="stylesheet/scss" lang="scss">
 //reset element css of el-icon-close
 .tags-view-wrapper {
-  .tags-view-item {
-    .el-icon-close {
-      width: 16px;
-      height: 16px;
-      vertical-align: 2px;
-      border-radius: 50%;
-      text-align: center;
-      transition: all .3s cubic-bezier(.645, .045, .355, 1);
-      transform-origin: 100% 50%;
-      &:before {
-        transform: scale(.6);
-        display: inline-block;
-        vertical-align: -3px;
-      }
-      &:hover {
-        background-color: #b4bccc;
-        color: #fff;
-      }
-    }
-  }
+	.tags-view-item {
+		.el-icon-close {
+			width: 16px;
+			height: 16px;
+			vertical-align: 2px;
+			border-radius: 50%;
+			text-align: center;
+			transition: all 0.3s cubic-bezier(0.645, 0.045, 0.355, 1);
+			transform-origin: 100% 50%;
+			&:before {
+				transform: scale(0.6);
+				display: inline-block;
+				vertical-align: -3px;
+			}
+			&:hover {
+				background-color: #b4bccc;
+				color: #fff;
+			}
+		}
+	}
 }
 </style>
