@@ -1,4 +1,4 @@
-export { default as Navbar } from './Navbar.vue';
-export { default as Sidebar } from './Sidebar/index.vue';
-export { default as AppMain } from './AppMain.vue';
-export { default as TagsView } from './TagsView.vue';
+export { default as Navbar } from "./Navbar.vue";
+export { default as Sidebar } from "./Sidebar/index.vue";
+export { default as AppMain } from "./AppMain.vue";
+export { default as TagsView } from "./TagsView.vue";
