@@ -1,4 +1,4 @@
-import store from '@/store';
+import store from "@/store";
 
 const { body } = document;
 const WIDTH = 768;
@@ -6,35 +6,35 @@ const RATIO = 3;
 
 export default {
 	watch: {
-		$route (route) {
-			if (this['device'] === 'mobile' && this['sidebar'].opened) {
-				store.dispatch('CloseSideBar', { withoutAnimation: false });
+		$route(route) {
+			if (this["device"] === "mobile" && this["sidebar"].opened) {
+				store.dispatch("CloseSideBar", { withoutAnimation: false });
 			}
-		}
+		},
 	},
-	beforeMount () {
-		window.addEventListener('resize', this['resizeHandler']);
+	beforeMount() {
+		window.addEventListener("resize", this["resizeHandler"]);
 	},
-	mounted () {
-		this['resizeHandler']();
+	mounted() {
+		this["resizeHandler"]();
 	},
 	methods: {
-		isMobile () {
+		isMobile() {
 			const rect = body.getBoundingClientRect();
 			return rect.width - RATIO < WIDTH;
 		},
-		resizeHandler () {
+		resizeHandler() {
 			if (!document.hidden) {
 				const isMobile = this.isMobile();
-				store.dispatch('ToggleDevice', isMobile ? 'mobile' : 'desktop');
+				store.dispatch("ToggleDevice", isMobile ? "mobile" : "desktop");
 
 				if (isMobile) {
-					store.dispatch('CloseSideBar', { withoutAnimation: true });
+					store.dispatch("CloseSideBar", { withoutAnimation: true });
 				} else {
 					const rect = body.getBoundingClientRect();
-					store.dispatch('OpenSideBar', { withoutAnimation: true });
+					store.dispatch("OpenSideBar", { withoutAnimation: true });
 				}
 			}
-		}
-	}
+		},
+	},
 };
