@@ -2,7 +2,7 @@
 	<div>
 		<el-card class="box-card">
 			<el-row type="flex" justify="space-between">
-				<div>{{type=='signin'?'登录':'注册'}}</div>
+				<div>{{type =='signin'?'登录':'注册'}}</div>
 				<!-- <toggle-mode></toggle-mode> -->
 			</el-row>
 			<br>
@@ -69,96 +69,118 @@
 	</div>
 </template>
 <script>
-import { setToken, removeToken } from '@/utils/auth';
-import { LOCAL_SESSION } from '@/.config';
-import { mapGetters } from 'vuex';
+import { setToken, removeToken } from "@/utils/auth";
+import { LOCAL_SESSION } from "@/.config";
+import { mapGetters } from "vuex";
 
 export default {
-	data () {
-		let validatePassword=  (rule, value, callback)=> {
-			if (value === '') {
-				callback(new Error('请再次输入密码'));
+	data() {
+		let validatePassword = (rule, value, callback) => {
+			if (value === "") {
+				callback(new Error("请再次输入密码"));
 			} else if (value !== this.registerForm.password) {
-				callback(new Error('两次输入密码不一致!'));
+				callback(new Error("两次输入密码不一致!"));
 			} else {
 				callback();
 			}
 		};
-		let validateAccount= (rule, value, callback)=> {
-			if(!value){
-				return callback();  
+		let validateAccount = (rule, value, callback) => {
+			if (!value) {
+				return callback();
 			}
 			if (/^[A-Za-z_]\w{2,30}$/.test(value)) {
-				return callback();  
+				return callback();
 			}
 			if (/^\d{11}$/.test(value)) {
-				return callback();  
+				return callback();
 			}
 			if (/^[\w.]{3,30}@[\w.]+$/.test(value)) {
-				return callback();  
+				return callback();
 			}
-			return callback(new Error('请输入正确的账号'));
+			return callback(new Error("请输入正确的账号"));
 		};
-		let validateName= (rule, value, callback)=> {
-			if(!value){
-				return callback();  
+		let validateName = (rule, value, callback) => {
+			if (!value) {
+				return callback();
 			}
 			if (/^[A-Za-z_]\w{2,30}$/.test(value)) {
-				return callback();  
+				return callback();
 			}
-			return callback(new Error('请输入正确的账号'));
+			return callback(new Error("请输入正确的账号"));
 		};
 		return {
-			type:'signin',
+			type: "signin",
 			form: {
 				login: null,
 				password: null,
-				client: 'ACC'
+				client: "ACC",
 			},
 			registerForm: {
 				username: null,
 				password: null,
-				repassword:null,
-				phone:undefined,
-				email:undefined,
-				client: 'ACC'
+				repassword: null,
+				phone: undefined,
+				email: undefined,
+				client: "ACC",
 			},
 			rememberMe: false,
 			loading: false,
 			rules: {
 				login: [
-					{ required: true, message: '请输入帐号', trigger: 'blur' },
-					{ min: 3, max: 30, message: '长度至少 3 个字符', trigger: 'blur' },
-					{ validator: validateAccount, trigger: 'blur' }
+					{ required: true, message: "请输入帐号", trigger: "blur" },
+					{
+						min: 3,
+						max: 30,
+						message: "长度至少 3 个字符",
+						trigger: "blur",
+					},
+					{ validator: validateAccount, trigger: "blur" },
 				],
 				password: [
-					{ required: true, message: '请输入密码', trigger: 'blur' },
-					{ min: 6, max: 30, message: '长度至少 6 个字符', trigger: 'blur' }
+					{ required: true, message: "请输入密码", trigger: "blur" },
+					{
+						min: 6,
+						max: 30,
+						message: "长度至少 6 个字符",
+						trigger: "blur",
+					},
 				],
 			},
 			registerRules: {
 				username: [
-					{ required: true, message: '请输入帐号', trigger: 'blur' },
-					{ min: 3, max: 30, message: '长度至少 3 个字符', trigger: 'blur' },
-					{ validator: validateName, trigger: 'blur' }
+					{ required: true, message: "请输入帐号", trigger: "blur" },
+					{
+						min: 3,
+						max: 30,
+						message: "长度至少 3 个字符",
+						trigger: "blur",
+					},
+					{ validator: validateName, trigger: "blur" },
 				],
 				password: [
-					{ required: true, message: '请输入密码', trigger: 'blur' },
-					{ min: 6, max: 30, message: '长度至少 6 个字符', trigger: 'blur' }
+					{ required: true, message: "请输入密码", trigger: "blur" },
+					{
+						min: 6,
+						max: 30,
+						message: "长度至少 6 个字符",
+						trigger: "blur",
+					},
 				],
-				repassword: [
-					{ validator: validatePassword, trigger: 'blur' }
-				],
-				email:[
-					{ type: 'email', message: '请输入正确的邮箱地址', trigger: ['blur', 'change'] }
+				repassword: [{ validator: validatePassword, trigger: "blur" }],
+				email: [
+					{
+						type: "email",
+						message: "请输入正确的邮箱地址",
+						trigger: ["blur", "change"],
+					},
 				],
 			},
 			redirect: null,
 			pwd: {
 				view: false,
-				icon: 'fa-eye-slash',
-				type: 'password'
-			}
+				icon: "fa-eye-slash",
+				type: "password",
+			},
 		};
 	},
 	components: {
@@ -166,129 +188,137 @@ export default {
 	},
 	watch: {
 		$route: {
-			handler (route) {
+			handler(route) {
 				this.redirect = route.query && route.query.redirect;
 			},
-			immediate: true
-		}
+			immediate: true,
+		},
 	},
 	computed: {
-		...mapGetters(['account'])
+		...mapGetters(["account"]),
 	},
-	mounted () {
+	mounted() {
 		if (this.account) {
 			this.form.login = this.account;
 			this.rememberMe = true;
 		}
-		this.$store.dispatch('checkLogin').then(() => {
+		this.$store.dispatch("checkLogin").then(() => {
 			if (this.$store.getters.isLogin) {
-				this.$confirm(`${this.$store.getters.user} 已登录，是否前往主页？`, '提示', {
-					confirmButtonText: '确定',
-					cancelButtonText: '重新登录',
-					type: 'success',
-				}).then(() => {
-					// 避免 浏览器前进后退 <- ->导致的页面错误
-					if (!this.$route.path.includes('login')) {
-						return;
+				this.$confirm(
+					`${this.$store.getters.user} 已登录，是否前往主页？`,
+					"提示",
+					{
+						confirmButtonText: "确定",
+						cancelButtonText: "重新登录",
+						type: "success",
 					}
-					this.$router.push({ path: this.redirect || '/' });
-				}).catch(action => {
-					// 避免 浏览器前进后退 <- ->导致的页面错误
-					if (!this.$route.path.includes('login')) {
-						return;
-					}
-					if (action === 'cancel') {
-						this.$message({
-							type: 'warning',
-							message: '请重新登录'
-						});
-						removeToken();
-					}
-				});
+				)
+					.then(() => {
+						// 避免 浏览器前进后退 <- ->导致的页面错误
+						if (!this.$route.path.includes("login")) {
+							return;
+						}
+						this.$router.push({ path: this.redirect || "/" });
+					})
+					.catch(action => {
+						// 避免 浏览器前进后退 <- ->导致的页面错误
+						if (!this.$route.path.includes("login")) {
+							return;
+						}
+						if (action === "cancel") {
+							this.$message({
+								type: "warning",
+								message: "请重新登录",
+							});
+							removeToken();
+						}
+					});
 			}
 		});
 	},
 	methods: {
-		signin () {
+		signin() {
 			this.$refs.form.validate(valid => {
 				if (valid) {
 					if (this.loading) {
 						return;
 					}
 					this.loading = true;
-					this.$store.commit('SET_ACCOUNT', this.rememberMe ? this.form.login : '');
-					this.post(
-						'dis/access-tokens',
-						this.form
-					).then(res => {
-						const self = this;
-						this.$message({
-							message: '登录成功',
-							type: 'success',
+					this.$store.commit(
+						"SET_ACCOUNT",
+						this.rememberMe ? this.form.login : ""
+					);
+					this.post("dis/access-tokens", this.form)
+						.then(res => {
+							const self = this;
+							this.$message({
+								message: "登录成功",
+								type: "success",
+							});
+							setToken(res[LOCAL_SESSION], res.expireAt);
+							this.$store.dispatch("GetProfile");
+							self.$router.push({ path: self.redirect || "/" });
+						})
+						.catch(() => {
+							this.loading = false;
 						});
-						setToken(res[LOCAL_SESSION], res.expireAt);
-						this.$store.dispatch('GetProfile');
-						self.$router.push({ path: self.redirect || '/' });
-					}).catch(() => {
-						this.loading = false;
-					});
 				}
 			});
 		},
-		togglePwd () {
+		togglePwd() {
 			this.pwd.view = !this.pwd.view;
 			if (this.pwd.view) {
-				this.pwd.icon = 'fa-eye';
-				this.pwd.type = 'text';
+				this.pwd.icon = "fa-eye";
+				this.pwd.type = "text";
 			} else {
-				this.pwd.icon = 'fa-eye-slash';
-				this.pwd.type = 'password';
+				this.pwd.icon = "fa-eye-slash";
+				this.pwd.type = "password";
 			}
 		},
-		signup () {
+		signup() {
 			this.$refs.registerForm.validate(valid => {
 				if (valid) {
 					if (this.loading) {
 						return;
 					}
 					this.loading = true;
-					this.post(
-						'dis/users',
-						this.registerForm
-					).then(res => {
-						const self = this;
-						this.$message({
-							message: '注册成功',
-							type: 'success',
- 							onClose () {
- 								self.$router.push({ path: self.redirect || '/' });
- 							}
+					this.post("dis/users", this.registerForm)
+						.then(res => {
+							const self = this;
+							this.$message({
+								message: "注册成功",
+								type: "success",
+								onClose() {
+									self.$router.push({
+										path: self.redirect || "/",
+									});
+								},
+							});
+							setToken(res[LOCAL_SESSION], res.expireAt);
+							this.$store.dispatch("GetProfile");
+						})
+						.catch(() => {
+							this.loading = false;
 						});
-						setToken(res[LOCAL_SESSION], res.expireAt);
-						this.$store.dispatch('GetProfile');
-					}).catch(() => {
-						this.loading = false;
-					});
 				}
 			});
 		},
-		swichType(type,ref){
-			this.type=type;
+		swichType(type, ref) {
+			this.type = type;
 			this.$refs[ref].clearValidate();
-		}
+		},
 	},
 };
 </script>
 <style lang="scss" scoped>
 .box-card {
-  margin: 100px auto;
-  max-width: 500px;
+	margin: 100px auto;
+	max-width: 500px;
 }
 @media screen and (max-width: 500px) {
-  .box-card {
-    margin-top: 0;
-    width: 100%;
-  }
+	.box-card {
+		margin-top: 0;
+		width: 100%;
+	}
 }
 </style>
-
